Refactor to avoid loading tensorflow into ChunkParser workers. (#173)
* Refactor to avoid loading tensorflow into ChunkParser workers.

* Fixes for experimental path

* Fix for validation without experimental.
