+++ tf/train.py
@@ -24,12 +24,9 @@ import glob
 import gzip
 import random
 import multiprocessing as mp
-import tensorflow as tf
-from tfprocess import TFProcess
 from chunkparser import ChunkParser
 
 SKIP = 32
-SKIP_MULTIPLE = 1024
 
 
 def get_chunks(data_prefix):
@@ -87,283 +84,26 @@ def game_number_for_name(name):
     return int(num_str)
 
 
-def extract_policy_bits(raw):
-    # Next 7432 are easy, policy extraction.
-    policy = tf.io.decode_raw(tf.strings.substr(raw, 8, 7432), tf.float32)
-    # Next are 104 bit packed chess boards, they have to be expanded.
-    bit_planes = tf.expand_dims(
-        tf.reshape(
-            tf.io.decode_raw(tf.strings.substr(raw, 7440, 832), tf.uint8),
-            [-1, 104, 8]), -1)
-    bit_planes = tf.bitwise.bitwise_and(tf.tile(bit_planes, [1, 1, 1, 8]),
-                                        [128, 64, 32, 16, 8, 4, 2, 1])
-    bit_planes = tf.minimum(1., tf.cast(bit_planes, tf.float32))
-    return policy, bit_planes
-
-
-def extract_byte_planes(raw):
-    # 5 bytes in input are expanded and tiled
-    unit_planes = tf.expand_dims(
-        tf.expand_dims(
-            tf.io.decode_raw(tf.strings.substr(raw, 8272, 5), tf.uint8), -1),
-        -1)
-    unit_planes = tf.tile(unit_planes, [1, 1, 8, 8])
-    return unit_planes
-
-
-def extract_rule50_zero_one(raw):
-    # rule50 count plane.
-    rule50_plane = tf.expand_dims(
-        tf.expand_dims(
-            tf.io.decode_raw(tf.strings.substr(raw, 8277, 1), tf.uint8), -1),
-        -1)
-    rule50_plane = tf.cast(tf.tile(rule50_plane, [1, 1, 8, 8]), tf.float32)
-    rule50_plane = tf.divide(rule50_plane, 99.)
-    # zero plane and one plane
-    zero_plane = tf.zeros_like(rule50_plane)
-    one_plane = tf.ones_like(rule50_plane)
-    return rule50_plane, zero_plane, one_plane
-
-
-def extract_rule50_100_zero_one(raw):
-    # rule50 count plane.
-    rule50_plane = tf.expand_dims(
-        tf.expand_dims(
-            tf.io.decode_raw(tf.strings.substr(raw, 8277, 1), tf.uint8), -1),
-        -1)
-    rule50_plane = tf.cast(tf.tile(rule50_plane, [1, 1, 8, 8]), tf.float32)
-    rule50_plane = tf.divide(rule50_plane, 100.)
-    # zero plane and one plane
-    zero_plane = tf.zeros_like(rule50_plane)
-    one_plane = tf.ones_like(rule50_plane)
-    return rule50_plane, zero_plane, one_plane
-
-
-def extract_invariance(raw):
-    # invariance plane.
-    invariance_plane = tf.expand_dims(
-        tf.expand_dims(
-            tf.io.decode_raw(tf.strings.substr(raw, 8278, 1), tf.uint8), -1),
-        -1)
-    return tf.cast(tf.tile(invariance_plane, [1, 1, 8, 8]), tf.float32)
-
-
-def extract_outputs(raw):
-    # winner is stored in one signed byte and needs to be converted to one hot.
-    winner = tf.cast(
-        tf.io.decode_raw(tf.strings.substr(raw, 8279, 1), tf.int8), tf.float32)
-    winner = tf.tile(winner, [1, 3])
-    z = tf.cast(tf.equal(winner, [1., 0., -1.]), tf.float32)
-
-    # Outcome distribution needs to be calculated from q and d.
-    best_q = tf.io.decode_raw(tf.strings.substr(raw, 8284, 4), tf.float32)
-    best_d = tf.io.decode_raw(tf.strings.substr(raw, 8292, 4), tf.float32)
-    best_q_w = 0.5 * (1.0 - best_d + best_q)
-    best_q_l = 0.5 * (1.0 - best_d - best_q)
-
-    q = tf.concat([best_q_w, best_d, best_q_l], 1)
-
-    ply_count = tf.io.decode_raw(tf.strings.substr(raw, 8304, 4), tf.float32)
-    return z, q, ply_count
-
-
-def extract_inputs_outputs_if1(raw):
-    # first 4 bytes in each batch entry are boring.
-    # Next 4 change how we construct some of the unit planes.
-    #input_format = tf.reshape(
-    #    tf.io.decode_raw(tf.strings.substr(raw, 4, 4), tf.int32),
-    #    [-1, 1, 1, 1])
-    # tf.debugging.assert_equal(input_format, tf.ones_like(input_format))
-
-    policy, bit_planes = extract_policy_bits(raw)
-
-    # Next 5 are castling + stm, all of which simply copy the byte value to all squares.
-    unit_planes = tf.cast(extract_byte_planes(raw), tf.float32)
-
-    rule50_plane, zero_plane, one_plane = extract_rule50_zero_one(raw)
-
-    inputs = tf.reshape(
-        tf.concat(
-            [bit_planes, unit_planes, rule50_plane, zero_plane, one_plane], 1),
-        [-1, 112, 64])
-
-    z, q, ply_count = extract_outputs(raw)
-
-    return (inputs, policy, z, q, ply_count)
-
-
-def extract_unit_planes_with_bitsplat(raw):
-    unit_planes = extract_byte_planes(raw)
-    bitsplat_unit_planes = tf.bitwise.bitwise_and(
-        unit_planes, [1, 2, 4, 8, 16, 32, 64, 128])
-    bitsplat_unit_planes = tf.minimum(
-        1., tf.cast(bitsplat_unit_planes, tf.float32))
-    unit_planes = tf.cast(unit_planes, tf.float32)
-    return unit_planes, bitsplat_unit_planes
-
-
-def make_frc_castling(bitsplat_unit_planes, zero_plane):
-    queenside = tf.concat([
-        bitsplat_unit_planes[:, :1, :1], zero_plane[:, :, :6],
-        bitsplat_unit_planes[:, 2:3, :1]
-    ], 2)
-    kingside = tf.concat([
-        bitsplat_unit_planes[:, 1:2, :1], zero_plane[:, :, :6],
-        bitsplat_unit_planes[:, 3:4, :1]
-    ], 2)
-    return queenside, kingside
-
-
-def extract_inputs_outputs_if2(raw):
-    # first 4 bytes in each batch entry are boring.
-    # Next 4 change how we construct some of the unit planes.
-    #input_format = tf.reshape(
-    #    tf.io.decode_raw(tf.strings.substr(raw, 4, 4), tf.int32),
-    #    [-1, 1, 1, 1])
-    # tf.debugging.assert_equal(input_format, tf.multiply(tf.ones_like(input_format), 2))
-
-    policy, bit_planes = extract_policy_bits(raw)
-
-    # Next 5 inputs are 4 frc castling and 1 stm.
-    # In order to do frc we need to make bit unpacked versions.  Note little endian for these fields so the bitwise_and array is reversed.
-    # Although we only need bit unpacked for first 4 of 5 planes, its simpler just to create them all.
-    unit_planes, bitsplat_unit_planes = extract_unit_planes_with_bitsplat(raw)
-
-    rule50_plane, zero_plane, one_plane = extract_rule50_zero_one(raw)
-
-    # For FRC the old unit planes must be replaced with 0 and 2 merged, 1 and 3 merged, two zero planes and then original 4.
-    queenside, kingside = make_frc_castling(bitsplat_unit_planes, zero_plane)
-    unit_planes = tf.concat(
-        [queenside, kingside, zero_plane, zero_plane, unit_planes[:, 4:]], 1)
-
-    inputs = tf.reshape(
-        tf.concat(
-            [bit_planes, unit_planes, rule50_plane, zero_plane, one_plane], 1),
-        [-1, 112, 64])
-
-    z, q, ply_count = extract_outputs(raw)
-
-    return (inputs, policy, z, q, ply_count)
-
-
-def make_canonical_unit_planes(bitsplat_unit_planes, zero_plane):
-    # For canonical the old unit planes must be replaced with 0 and 2 merged, 1 and 3 merged, two zero planes and then en-passant.
-    queenside, kingside = make_frc_castling(bitsplat_unit_planes, zero_plane)
-    enpassant = tf.concat(
-        [zero_plane[:, :, :7], bitsplat_unit_planes[:, 4:, :1]], 2)
-    unit_planes = tf.concat(
-        [queenside, kingside, zero_plane, zero_plane, enpassant], 1)
-    return unit_planes
-
-
-def extract_inputs_outputs_if3(raw):
-    # first 4 bytes in each batch entry are boring.
-    # Next 4 change how we construct some of the unit planes.
-    #input_format = tf.reshape(
-    #    tf.io.decode_raw(tf.strings.substr(raw, 4, 4), tf.int32),
-    #    [-1, 1, 1, 1])
-    # tf.debugging.assert_equal(input_format, tf.multiply(tf.ones_like(input_format), 3))
-
-    policy, bit_planes = extract_policy_bits(raw)
-
-    # Next 5 inputs are 4 castling and 1 enpassant.
-    # In order to do the frc castling and if3 enpassant plane we need to make bit unpacked versions.  Note little endian for these fields so the bitwise_and array is reversed.
-    unit_planes, bitsplat_unit_planes = extract_unit_planes_with_bitsplat(raw)
-
-    rule50_plane, zero_plane, one_plane = extract_rule50_zero_one(raw)
-
-    unit_planes = make_canonical_unit_planes(bitsplat_unit_planes, zero_plane)
-
-    inputs = tf.reshape(
-        tf.concat(
-            [bit_planes, unit_planes, rule50_plane, zero_plane, one_plane], 1),
-        [-1, 112, 64])
-
-    z, q, ply_count = extract_outputs(raw)
-
-    return (inputs, policy, z, q, ply_count)
-
-
-def extract_inputs_outputs_if4(raw):
-    # first 4 bytes in each batch entry are boring.
-    # Next 4 change how we construct some of the unit planes.
-    #input_format = tf.reshape(
-    #    tf.io.decode_raw(tf.strings.substr(raw, 4, 4), tf.int32),
-    #    [-1, 1, 1, 1])
-    # tf.debugging.assert_equal(input_format, tf.multiply(tf.ones_like(input_format), 3))
-
-    policy, bit_planes = extract_policy_bits(raw)
-
-    # Next 5 inputs are 4 castling and 1 enpassant.
-    # In order to do the frc castling and if3 enpassant plane we need to make bit unpacked versions.  Note little endian for these fields so the bitwise_and array is reversed.
-    unit_planes, bitsplat_unit_planes = extract_unit_planes_with_bitsplat(raw)
-
-    rule50_plane, zero_plane, one_plane = extract_rule50_100_zero_one(raw)
-
-    unit_planes = make_canonical_unit_planes(bitsplat_unit_planes, zero_plane)
-
-    inputs = tf.reshape(
-        tf.concat(
-            [bit_planes, unit_planes, rule50_plane, zero_plane, one_plane], 1),
-        [-1, 112, 64])
-
-    z, q, ply_count = extract_outputs(raw)
-
-    return (inputs, policy, z, q, ply_count)
-
-
-def make_armageddon_stm(invariance_plane):
-    # invariance_plane contains values of 128 or higher if its black side to move, 127 or lower otherwise.
-    # Convert this to 0,1 by subtracting off 127 and then clipping.
-    return tf.clip_by_value(invariance_plane - 127., 0., 1.)
-
-
-def extract_inputs_outputs_if132(raw):
-    # first 4 bytes in each batch entry are boring.
-    # Next 4 change how we construct some of the unit planes.
-    #input_format = tf.reshape(
-    #    tf.io.decode_raw(tf.strings.substr(raw, 4, 4), tf.int32),
-    #    [-1, 1, 1, 1])
-    # tf.debugging.assert_equal(input_format, tf.multiply(tf.ones_like(input_format), 3))
-
-    policy, bit_planes = extract_policy_bits(raw)
-
-    # Next 5 inputs are 4 castling and 1 enpassant.
-    # In order to do the frc castling and if3 enpassant plane we need to make bit unpacked versions.  Note little endian for these fields so the bitwise_and array is reversed.
-    unit_planes, bitsplat_unit_planes = extract_unit_planes_with_bitsplat(raw)
-
-    rule50_plane, zero_plane, one_plane = extract_rule50_100_zero_one(raw)
-
-    unit_planes = make_canonical_unit_planes(bitsplat_unit_planes, zero_plane)
-
-    armageddon_stm = make_armageddon_stm(extract_invariance(raw))
-
-    inputs = tf.reshape(
-        tf.concat(
-            [bit_planes, unit_planes, rule50_plane, armageddon_stm, one_plane],
-            1), [-1, 112, 64])
-
-    z, q, ply_count = extract_outputs(raw)
-
-    return (inputs, policy, z, q, ply_count)
-
-
-def select_extractor(mode):
-    if mode == 1:
-        return extract_inputs_outputs_if1
-    if mode == 2:
-        return extract_inputs_outputs_if2
-    if mode == 3:
-        return extract_inputs_outputs_if3
-    if mode == 4 or mode == 5:
-        return extract_inputs_outputs_if4
-    if mode == 132 or mode == 133:
-        return extract_inputs_outputs_if132
-    assert (false)
-
-
-def semi_sample(x):
-    return tf.slice(tf.random.shuffle(x), [0], [SKIP_MULTIPLE])
+def get_input_mode(cfg):
+    import proto.net_pb2 as pb
+    input_mode = cfg['model'].get('input_type', 'classic')
+
+    if input_mode == "classic":
+        return pb.NetworkFormat.INPUT_CLASSICAL_112_PLANE
+    elif input_mode == "frc_castling":
+        return pb.NetworkFormat.INPUT_112_WITH_CASTLING_PLANE
+    elif input_mode == "canonical":
+        return pb.NetworkFormat.INPUT_112_WITH_CANONICALIZATION
+    elif input_mode == "canonical_100":
+        return pb.NetworkFormat.INPUT_112_WITH_CANONICALIZATION_HECTOPLIES
+    elif input_mode == "canonical_armageddon":
+        return pb.NetworkFormat.INPUT_112_WITH_CANONICALIZATION_HECTOPLIES_ARMAGEDDON
+    elif input_mode == "canonical_v2":
+        return pb.NetworkFormat.INPUT_112_WITH_CANONICALIZATION_V2
+    elif input_mode == "canonical_v2_armageddon":
+        return pb.NetworkFormat.INPUT_112_WITH_CANONICALIZATION_V2_ARMAGEDDON
+    else:
+        raise ValueError("Unknown input mode format: {}".format(input_mode))
 
 
 def main(cmd):
@@ -419,12 +159,9 @@ def main(cmd):
     root_dir = os.path.join(cfg['training']['path'], cfg['name'])
     if not os.path.exists(root_dir):
         os.makedirs(root_dir)
-    tfprocess = TFProcess(cfg)
     experimental_reads = max(2, mp.cpu_count() - 2) // 2
-    extractor = select_extractor(tfprocess.INPUT_MODE)
 
-    if experimental_parser and (diff_focus_min != 1
-                                or diff_focus_slope != 0):
+    if experimental_parser and (diff_focus_min != 1 or diff_focus_slope != 0):
         raise ValueError(
             'Experimental parser does not support non-default value \
                           focus parameters.')
@@ -436,15 +173,29 @@ def main(cmd):
             compression_type='GZIP',
             num_parallel_reads=experimental_reads)
 
+    test_shuffle_size = int(shuffle_size * (1.0 - train_ratio))
+
     if experimental_parser:
+        import tensorflow as tf
+        from tfprocess import TFProcess
+        tfprocess = TFProcess(cfg)
+        from experimental_parsing import select_extractor
+        from experimental_parsing import semi_sample
+        from experimental_parsing import SKIP_MULTIPLE
+        extractor = select_extractor(get_input_mode(cfg))
         train_dataset = tf.data.Dataset.from_tensor_slices(train_chunks).shuffle(len(train_chunks)).repeat().batch(256)\
                          .interleave(read, num_parallel_calls=2)\
                          .batch(SKIP_MULTIPLE*SKIP).map(semi_sample).unbatch()\
                          .shuffle(shuffle_size)\
                          .batch(split_batch_size).map(extractor)
+        test_dataset = tf.data.Dataset.from_tensor_slices(test_chunks).shuffle(len(test_chunks)).repeat().batch(256)\
+                         .interleave(read, num_parallel_calls=2)\
+                         .batch(SKIP_MULTIPLE*SKIP).map(semi_sample).unbatch()\
+                         .shuffle(test_shuffle_size)\
+                         .batch(split_batch_size).map(extractor)
     else:
         train_parser = ChunkParser(train_chunks,
-                                   tfprocess.INPUT_MODE,
+                                   get_input_mode(cfg),
                                    shuffle_size=shuffle_size,
                                    sample=SKIP,
                                    batch_size=ChunkParser.BATCH_SIZE,
@@ -453,34 +204,32 @@ def main(cmd):
                                    diff_focus_q_weight=diff_focus_q_weight,
                                    diff_focus_pol_scale=diff_focus_pol_scale,
                                    workers=train_workers)
+        # no diff focus for test_parser
+        test_parser = ChunkParser(test_chunks,
+                                  get_input_mode(cfg),
+                                  shuffle_size=test_shuffle_size,
+                                  sample=SKIP,
+                                  batch_size=ChunkParser.BATCH_SIZE,
+                                  workers=test_workers)
+        import tensorflow as tf
+        from chunkparsefunc import parse_function
+        from tfprocess import TFProcess
+        tfprocess = TFProcess(cfg)
         train_dataset = tf.data.Dataset.from_generator(
             train_parser.parse,
             output_types=(tf.string, tf.string, tf.string, tf.string,
                           tf.string))
-        train_dataset = train_dataset.map(ChunkParser.parse_function)
-
-    shuffle_size = int(shuffle_size * (1.0 - train_ratio))
-    if experimental_parser:
-        test_dataset = tf.data.Dataset.from_tensor_slices(test_chunks).shuffle(len(test_chunks)).repeat().batch(256)\
-                         .interleave(read, num_parallel_calls=2)\
-                         .batch(SKIP_MULTIPLE*SKIP).map(semi_sample).unbatch()\
-                         .shuffle(shuffle_size)\
-                         .batch(split_batch_size).map(extractor)
-    else:
-        # no diff focus for test_parser
-        test_parser = ChunkParser(test_chunks,
-                                  tfprocess.INPUT_MODE,
-                                  shuffle_size=shuffle_size,
-                                  sample=SKIP,
-                                  batch_size=ChunkParser.BATCH_SIZE,
-                                  workers=test_workers)
+        train_dataset = train_dataset.map(parse_function)
         test_dataset = tf.data.Dataset.from_generator(
             test_parser.parse,
             output_types=(tf.string, tf.string, tf.string, tf.string,
                           tf.string))
-        test_dataset = test_dataset.map(ChunkParser.parse_function)
+        test_dataset = test_dataset.map(parse_function)
+
     validation_dataset = None
     if 'input_validation' in cfg['dataset']:
+        from experimental_parsing import select_extractor
+        extractor = select_extractor(get_input_mode(cfg))
         valid_chunks = get_all_chunks(cfg['dataset']['input_validation'])
         validation_dataset = tf.data.FixedLengthRecordDataset(valid_chunks, 8308, compression_type='GZIP', num_parallel_reads=experimental_reads)\
                                .batch(split_batch_size, drop_remainder=True).map(extractor)
